change script locate
