--- gomi/Assets/NewBehaviourScript.cs
@@ -1,18 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class NewBehaviourScript : MonoBehaviour
-{
-
-    // Start is called before the first frame update
-    void Start()
-    {
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        transform.position += new Vector3(0f, 0f, -0.05f);
-    }
-}
