+++ gomi/Assets/Scripts/NewBehaviourScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 02b2cc53117c84725b5a8b5b8fe0f7d1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
